requirements.txt
pgvector
django
numpy
sentence-transformers
openai
psycopg2-binary
martor
django-tables2
django-filter
